+++ src/main/java/Clases_Dominio/Cronometro.java
@@ -1,3 +1,4 @@
+package Clases_Dominio;
 
 public class Cronometro extends Thread {
