+++ src/main/java/Clases_Dominio/Persona.java
@@ -2,6 +2,12 @@ package Clases_Dominio;
 
 public class Persona {
 
+    /*
+    creo que esta clase solo debe  tener datos de persona (solo el nick),
+    en otra clase guardar las estadisticas de cada persona en un List<Persona> persona = new ...
+    ya que en esta  se guardarian los datos referentes a cada partidad que juegue una persona
+     */
+
     private String nick;
 
     public Persona(String nick) {
