+++ src/main/java/Clases_Dominio/Validaciones.java
@@ -1,3 +1,5 @@
+package Clases_Dominio;
+
 import java.util.Scanner;
 
 public class Validaciones {
