+++ src/test/java/MemoriceTest.java
@@ -1,5 +1,4 @@
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
+import Clases_Dominio.Memorice;
 
 class MemoriceTest {
